Move tests to test folder, funcs to highlight folder
