+++ salve_ipc/server_functions/highlight/tree_sitter_funcs.py
